+++ src/Components/Video/OneVideo.js
@@ -1,7 +1,7 @@
 import React, {useState} from "react";
 import {useNavigate} from "react-router-dom";
 // Components
-import PlaylistToggle from "../Playlist/PlaylistToggle";
+import AddVideoToPlaylistToggle from "../Playlist/AddVideoToPlaylistToggle";
 // Style
 import styled from "styled-components";
 import StyledBtn from "../../Style/StyledBtn";
@@ -54,7 +54,7 @@ const OneVideo = ({video}) => {
                   onClick={onClickAddToPlaylist}
                   style={{marginLeft: "4px", cursor: "pointer"}}/>
               {clickedPlyId === videoId &&   /*클릭한 아이콘과 id가 동일한 모달창에만 적용되도록*/
-                  <PlaylistToggle
+                  <AddVideoToPlaylistToggle
                       show={playListToggleDisplay}
                       setShow={setPlayListToggleDisplay}
                   />
