+++ src/Pages/Playlist/PlaylistDetail.js
@@ -24,6 +24,7 @@ import {
 import Dot3_btn from "../../Assets/Dot3_btn.png";
 import Stored_Ply from "../../Assets/Stored_Ply.png";
 import Add_New_Ply from "../../Assets/Add_New_Ply.png";
+import {PlaylistHeader} from "./AllPlaylist";
 
 
 
@@ -138,12 +139,7 @@ const PlaylistDetail = () => {
   return (
       <div>
         <Header />
-        <TopBtns>
-          <Link to="/playlist/subscribe">
-            <img src={Stored_Ply} alt="구독 플레이리스트 보기" />
-          </Link>
-          <img src={Add_New_Ply} alt="새 플레이리스트 만들기" />
-        </TopBtns>
+          <PlaylistHeader />
         <Wrapper>
           <TitleWrapper>
             <DetailTitle>
